+++ src/components/index.ts
@@ -5,4 +5,3 @@ export { default as FloatingElements } from './FloatingElements';
 export { default as GradientText } from './GradientText';
 export { default as Grid } from './Grid';
 export { default as Section } from './Section';
-